requirements.txt
flask>=3.0
apscheduler>=3.10
requests>=2.31
pyyaml>=6.0
jinja2>=3.1
gunicorn>=22.0
